validate & invoke cli builds
