+++ internal/collector/initializer/metric.go
@@ -2,7 +2,6 @@ package initializer
 
 import (
 	"fmt"
-	"github.com/turbolytics/latte/internal/collector/config"
 	"github.com/turbolytics/latte/internal/collector/metric"
 	"github.com/turbolytics/latte/internal/invoker"
 	"github.com/turbolytics/latte/internal/schedule"
@@ -14,9 +13,6 @@ import (
 	"github.com/turbolytics/latte/internal/source"
 	"github.com/turbolytics/latte/internal/state"
 	"go.uber.org/zap"
-	"gopkg.in/yaml.v3"
-	"os"
-	"path/filepath"
 )
 
 func NewStorer(c state.Config, l *zap.Logger) (invoker.Storer, error) {
@@ -33,87 +29,6 @@ func NewStorer(c state.Config, l *zap.Logger) (invoker.Storer, error) {
 		return nil, err
 	}
 	return s, err
-}
-
-type RootOption func(*config.Root)
-
-func WithJustValidation(validate bool) RootOption {
-	return func(c *config.Root) {
-		c.Validate = validate
-	}
-}
-
-func RootWithLogger(l *zap.Logger) RootOption {
-	return func(c *config.Root) {
-		c.Logger = l
-	}
-}
-
-func NewCollectorsFromGlob(glob string, opts ...RootOption) ([]invoker.Collector, error) {
-	files, err := filepath.Glob(glob)
-	if err != nil {
-		return nil, err
-	}
-	var collectors []invoker.Collector
-
-	for _, fName := range files {
-		c, err := NewCollectorFromFile(fName, opts...)
-		if err != nil {
-			return nil, err
-		}
-		collectors = append(collectors, c)
-	}
-	return collectors, nil
-}
-
-func NewCollectorFromFile(fpath string, opts ...RootOption) (invoker.Collector, error) {
-	fmt.Printf("loading config from file: %q\n", fpath)
-
-	bs, err := os.ReadFile(fpath)
-	if err != nil {
-		return nil, err
-	}
-	return NewCollector(bs, opts...)
-}
-
-func NewCollector(bs []byte, opts ...RootOption) (invoker.Collector, error) {
-	var conf config.Root
-
-	for _, opt := range opts {
-		opt(&conf)
-	}
-
-	if err := yaml.Unmarshal(bs, &conf); err != nil {
-		return nil, err
-	}
-
-	var coll invoker.Collector
-	var err error
-
-	switch conf.Collector.Type {
-	case config.TypeMetric:
-		coll, err = NewMetricCollectorFromConfig(
-			bs,
-			conf.Validate,
-			conf.Logger,
-		)
-	case config.TypePartition:
-		/*
-			collConfig, err := partition.NewConfig(
-				bs,
-				partition.ConfigWithJustValidation(conf.validate),
-				partition.ConfigWithLogger(conf.logger),
-			)
-		*/
-	default:
-		return nil, fmt.Errorf("collector type: %v not supported", conf.Collector.Type)
-	}
-
-	if err != nil {
-		return nil, err
-	}
-
-	return coll, err
 }
 
 func NewSink(c sink.Config, l *zap.Logger, validate bool) (invoker.Sinker, error) {
@@ -233,7 +148,6 @@ func NewMetricCollectorFromConfig(bs []byte, validate bool, l *zap.Logger) (*met
 		metric.WithSinks(sinks),
 		metric.WithSourcer(sourcer),
 		metric.WithStateStore(stateStore),
-
 	)
 
 	return coll, nil
