+++ internal/source/source.go
@@ -13,26 +13,3 @@ type Config struct {
 	Config map[string]any
 	Type   Type
 }
-
-func (c Config) Validate() error {
-	/*
-		vs := map[TypeStrategy]struct{}{
-			TypeStrategyTick:                   {},
-			TypeStrategyHistoricTumblingWindow: {},
-			TypeStrategyIncremental:            {},
-		}
-
-		if _, ok := vs[c.Strategy]; !ok {
-			return fmt.Errorf("unknown strategy: %q", c.Strategy)
-		}
-	*/
-	return nil
-}
-
-func (c *Config) SetDefaults() {
-	/*
-		if c.Strategy == "" {
-			c.Strategy = TypeStrategyTick
-		}
-	*/
-}
